requirements.txt
growwapi
streamlit
pandas
numpy
plotly
python-dotenv
